move PeriodicTimer class to separate file
